refactor: 현 구조에 따라 VPC 파일 위치 변경
