+++ src/main/java/com/agenticcp/core/domain/cloud/port/model/vpc/CreateVpcCommand.java
@@ -1,4 +1,6 @@
-package com.agenticcp.core.domain.cloud.port.command.vpc;
+package com.agenticcp.core.domain.cloud.port.model.vpc;
+
+import java.util.Map;
 
 import com.agenticcp.core.domain.cloud.entity.CloudProvider;
 import com.agenticcp.core.domain.cloud.port.model.account.CloudSessionCredential;
@@ -6,17 +8,21 @@ import com.agenticcp.core.domain.cloud.port.model.account.CloudSessionCredential
 import lombok.Builder;
 
 /**
- * VPC 삭제 명령.
+ * VPC 생성 시 아웃바운드 포트로 전달되는 명령 모델.
  */
 @Builder
-public record DeleteVpcCommand(
+public record CreateVpcCommand(
         CloudProvider.ProviderType providerType,
         String accountScope,
         String region,
-        String providerResourceId,
         String serviceKey,
         String resourceType,
+        String vpcName,
+        String cidrBlock,
+        String description,
+        Map<String, String> tags,
         String tenantKey,
+        Map<String, Object> providerSpecificConfig,
         CloudSessionCredential session
 ) {
 }
